Конструктор - тип поля "Файл" v2.2 rename
