+++ l2-frontend/src/construct/ParaclinicResearchEditorComponents/types/ParaclinicResearchEditor.ts
